Métodos para eliminar de las listas la id de un juego

+++ src/main/java/com/info/GamingList/model/Usuario.java
@@ -31,7 +31,7 @@ public class Usuario {
 
     public void eliminarIdLista(int id){
         if(idJuegos.contains(id)) {
-            idJuegos.remove(id);
+            idJuegos.remove((Integer) id);
         }
     }
 }
